Large-scale, WIP, command refactoring
Thesis got in the way of doing this, so I've sorta forgotten what I was actually doing with this.  Argh!

+++ URY.BAPS.Client.Common/Controllers/ChannelControllerSet.cs
@@ -16,7 +16,7 @@ namespace URY.BAPS.Client.Common.Controllers
         }
 
         [Pure]
-        protected override ChannelController MakeController(ushort channelId)
+        protected override ChannelController MakeController(byte channelId)
         {
             return new ChannelController(channelId, Core, _config);
         }
